+++ packages/protocol/src/cluster/client/ClusterClient.ts
@@ -91,6 +91,20 @@ export function ClusterClient<const T extends ClusterType>(
                 throw e;
             }
         };
+        result[`get${capitalizedAttributeName}AttributeFromCache`] = async () => {
+            if (isGroupAddress) {
+                throw new ImplementationError("Group cluster clients do not support reading attributes");
+            }
+
+            try {
+                return await (attributes as any)[attributeName].getLocal();
+            } catch (e) {
+                if (StatusResponseError.is(e, StatusCode.UnsupportedAttribute)) {
+                    return undefined;
+                }
+                throw e;
+            }
+        };
         result[`set${capitalizedAttributeName}Attribute`] = async <T>(value: T, dataVersion?: number) =>
             (attributes as any)[attributeName].set(value, dataVersion);
         result[`subscribe${capitalizedAttributeName}Attribute`] = async <T>(